# site/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Run402 API",
    "version": "2.0.0",
    "description": "Full-stack infrastructure for AI agents. Postgres, REST API, auth, file storage, static site hosting, serverless functions. Pay-per-tier via x402 USDC on Base — subscribe once, then use wallet signatures for all operations within your tier.",
    "contact": {
      "name": "Run402",
      "url": "https://run402.com"
    },
...
  },
  "servers": [
    {
      "url": "https://api.run402.com",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/deployments/v1": {
      "get": {
        "tags": [
          "Deployments"
        ],
        "summary": "Deployment info",
        "operationId": "deploymentsInfo",
        "description": "Returns deployment info. No auth required.",
        "responses": {
          "200": {
            "description": "Deployment info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "description": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "body": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Deployments"
        ],
        "summary": "Deploy a static site",
        "operationId": "createDeployment",
        "description": "Deploys static files to a live URL at {id}.sites.run402.com. Supports HTML, CSS, JS, images. SPA-ready with automatic fallback routing. Max 50MB total. Idempotent with Idempotency-Key header. Requires wallet auth, free within tier.",
        "security": [
          {
            "walletAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/WalletHeader"
          },
          {
            "$ref": "#/components/parameters/SignatureHeader"
          },
          {
            "$ref": "#/components/parameters/TimestampHeader"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "files"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Deployment name"
                  },
                  "project": {
                    "type": "string",
                    "description": "Optional project ID to associate with"
                  },
                  "target": {
                    "type": "string",
                    "description": "Optional deploy target"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "file",
                        "data"
                      ],
                      "properties": {
                        "file": {
                          "type": "string",
                          "description": "File path (e.g. index.html, css/style.css)"
                        },
                        "data": {
                          "type": "string",
                          "description": "File content (UTF-8 text or base64-encoded)"
                        },
                        "encoding": {
                          "type": "string",
                          "enum": [
                            "utf-8",
                            "base64"
                          ],
                          "default": "utf-8"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Site deployed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "status": {
                      "type": "string"
                    },
                    "files_count": {
                      "type": "integer"
                    },
                    "total_size": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/deployments/v1/{id}": {
      "get": {
        "tags": [
          "Deployments"
        ],
        "summary": "Get deployment status",
        "operationId": "getDeployment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deployment details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "status": {
                      "type": "string"
                    },
                    "files_count": {
                      "type": "integer"
                    },
                    "total_size": {
                      "type": "integer"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Deployment not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apikey": {
        "type": "apiKey",
        "in": "header",
        "name": "apikey",
        "description": "Project anon_key (JWT). Returned when you provision a project."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "User access_token from /auth/v1/token. Pass as Authorization: Bearer <token>."
      },
      "serviceKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Project service_key (JWT with service_role). Returned when you provision a project. Pass as Authorization: Bearer <service_key>."
      },
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Server admin key. Required for admin operations."
      },
      "x402": {
        "type": "apiKey",
        "in": "header",
        "name": "X-PAYMENT",
        "description": "x402 payment header. The client library (e.g. @anthropic-ai/sdk x402 facilitator) handles this automatically. Payment is in USDC on Base (mainnet eip155:8453 or testnet eip155:84532)."
      },
      "walletAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Run402-Wallet",
        "description": "EIP-4361 wallet authentication. Requires three headers: X-Run402-Wallet (wallet address), X-Run402-Signature (signed message), X-Run402-Timestamp (unix timestamp). The wallet must have an active tier subscription."
      }
    },
    "parameters": {
      "WalletHeader": {
        "name": "X-Run402-Wallet",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Ethereum wallet address (0x...)"
      },
      "SignatureHeader": {
        "name": "X-Run402-Signature",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "EIP-4361 signature of the message 'run402:{method}:{path}:{timestamp}'"
      },
      "TimestampHeader": {
        "name": "X-Run402-Timestamp",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Unix timestamp (must be within 30 seconds of server time)"
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid authentication. For wallet-auth endpoints, provide X-Run402-Wallet, X-Run402-Signature, and X-Run402-Timestamp headers. For project endpoints, provide apikey header (anon_key) or Authorization: Bearer (service_key/access_token).",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}